public ui: fix menu username
* Builds the menu during the first request.
* Store the logged user name in the session.
* Makes the translations and the logged user dynamic.
* Puts the menu generation code in a specific file.
* Closes: #2168.
