+++ rero_ils/theme/views.py
@@ -21,21 +21,17 @@ from __future__ import absolute_import, print_function
 
 import copy
 import re
-from functools import partial, wraps
+from functools import wraps
 
 from flask import Blueprint, abort, current_app, jsonify, redirect, \
     render_template, request, session, url_for
-from flask_babelex import gettext as _
-from flask_login import current_user
-from flask_menu import current_menu
-from invenio_i18n.ext import current_i18n
 from invenio_jsonschemas import current_jsonschemas
 from invenio_jsonschemas.errors import JSONSchemaNotFound
 
 from rero_ils.modules.organisations.api import Organisation
-from rero_ils.modules.patrons.api import current_librarian, current_patrons
 
-from ..permissions import admin_permission, can_access_professional_view
+from .menus import init_menu_lang, init_menu_profile, init_menu_tools
+from ..permissions import can_access_professional_view
 from ..version import __version__
 
 blueprint = Blueprint(
@@ -46,306 +42,7 @@ blueprint = Blueprint(
 )
 
 
-def rero_register(
-    item,
-    endpoint=None,
-    text=None,
-    order=0,
-    external_url=None,
-    endpoint_arguments_constructor=None,
-    dynamic_list_constructor=None,
-    active_when=None,
-    visible_when=None,
-    expected_args=None,
-        **kwargs):
-    """Take care each element in kwargs doesn't already exists in item."""
-    # Check which option in kwargs already exists in `item`.
-    to_delete = []
-    for option in kwargs.keys():
-        if hasattr(item, option):
-            to_delete.append(option)
-    # Delete all existing options in kwargs
-    for element in to_delete:
-        del kwargs[element]
-
-    return item.register(
-        endpoint,
-        text,
-        order,
-        external_url,
-        endpoint_arguments_constructor,
-        dynamic_list_constructor,
-        active_when,
-        visible_when,
-        expected_args,
-        **kwargs)
-
-
-def init_menu_tools():
-    """Create the header tool menu."""
-    item = current_menu.submenu('main.tool')
-    rero_register(
-        item,
-        endpoint=None,
-        text='{icon} <span class="{visible}">{menu}'.format(
-            icon='<i class="fa fa-wrench"></i>',
-            visible='visible-md-inline visible-lg-inline',
-            menu=_('Tools')
-        ),
-        order=0,
-        id='tools-menu'
-    )
-
-    item = current_menu.submenu('main.tool.ill_request')
-    rero_register(
-        item,
-        endpoint='ill_requests.ill_request_form',
-        visible_when=lambda: bool(current_patrons),
-        text='{icon} {help}'.format(
-            icon='<i class="fa fa-shopping-basket"></i>',
-            help=_('Interlibrary loan request')
-        ),
-        order=10,
-        id='ill-request-menu'
-    )
-
-    item = current_menu.submenu('main.tool.stats')
-    rero_register(
-        item,
-        endpoint='stats.stats',
-        visible_when=lambda: admin_permission.require().can(),
-        text='{icon} {help}'.format(
-            icon='<i class="fa fa-money"></i>',
-            help=_('Statistics')
-        ),
-        order=20,
-        id='stats-menu'
-    )
-
-    item = current_menu.submenu('main.tool.collections')
-    rero_register(
-        item,
-        endpoint='rero_ils.search',
-        endpoint_arguments_constructor=lambda: dict(
-            viewcode=request.view_args.get(
-                'viewcode', current_app.config.get(
-                    'RERO_ILS_SEARCH_GLOBAL_VIEW_CODE')),
-            recordType='collections'
-        ),
-        visible_when=lambda: current_app.config.get(
-            'RERO_ILS_SEARCH_GLOBAL_VIEW_CODE'
-            ) != request.view_args.get(
-                'viewcode', current_app.config.get(
-                    'RERO_ILS_SEARCH_GLOBAL_VIEW_CODE')),
-        text='{icon} {help}'.format(
-            icon='<i class="fa fa-graduation-cap"></i>',
-            help=_('Exhibition/course')
-        ),
-        order=2,
-        id='collections-menu'
-    )
-
-    item = current_menu.submenu('main.tool.help')
-    rero_register(
-        item,
-        endpoint='wiki.page',
-        endpoint_arguments_constructor=lambda: {'url': 'public'},
-        text='{icon} {help}'.format(
-            icon='<i class="fa fa-info"></i>',
-            help=_('Help')
-        ),
-        order=100,
-        id='help-menu'
-    )
-
-
-def init_menu_lang():
-    """Create the header language menu."""
-    item = current_menu.submenu('main.menu')
-    ui_language = 'ui_language_{lang}'.format(lang=current_i18n.language)
-    # Bug: when you reload the page with register(**kwargs), it failed
-    # We so check that 'id' already exists. If yes, do not create again
-    # the item.
-    rero_register(
-        item,
-        endpoint=None,
-        text='{icon} <span class="{visible}">{menu}'.format(
-            icon='<i class="fa fa-language"></i>',
-            visible='visible-md-inline visible-lg-inline',
-            menu=_(ui_language)
-        ),
-        order=0,
-        id='language-menu'
-    )
-
-    order = 10
-
-    def return_language(lang):
-        return dict(lang_code=lang)
-
-    def hide_language(lang):
-        return current_i18n.language != lang
-
-    for language_item in current_i18n.get_locales():
-        item = current_menu.submenu(
-            'main.menu.lang_{language}'.format(
-                language=language_item.language))
-        ui_language = 'ui_language_{lang}'.format(lang=language_item.language)
-        rero_register(
-            item,
-            endpoint='invenio_i18n.set_lang',
-            endpoint_arguments_constructor=partial(
-                return_language, language_item.language),
-            text='{icon} {language}'.format(
-                icon='<i class="fa fa-language"></i>',
-                language=_(ui_language)
-            ),
-            visible_when=partial(hide_language, language_item.language),
-            order=order,
-            id='language-menu-{language}'.format(
-                language=language_item.language))
-        order += 1
-
-
-def init_menu_profile():
-    """Create the profile header menu."""
-    item = current_menu.submenu('main.profile')
-    if current_librarian:
-        session['user_name'] = current_librarian.formatted_name
-    elif len(current_patrons) > 0:
-        session['user_name'] = current_patrons[0].formatted_name
-    else:
-        try:
-            session['user_name'] = current_user.email
-        # AnonymousUser
-        except AttributeError:
-            session.pop('user_name', None)
-    account = session.get('user_name', _('My Account'))
-    if len(account) > 30:
-        account = account[:30] + '…'
-
-    rero_register(
-        item,
-        endpoint=None,
-        text='<span class="btn btn-sm btn-success">{icon} '
-             '<span class="{visible}">{account}</span><span>'.format(
-                 icon='<i class="fa fa-user"></i>',
-                 visible='visible-md-inline visible-lg-inline',
-                 account=account
-             ),
-        order=1,
-        id='my-account-menu',
-        cssClass='py-1'
-    )
-
-    item = current_menu.submenu('main.profile.login')
-
-    rero_register(
-        item,
-        endpoint='security.login',
-        endpoint_arguments_constructor=lambda: dict(
-            next=request.full_path
-        ),
-        visible_when=lambda: not current_user.is_authenticated,
-        text='{icon} {login}'.format(
-            icon='<i class="fa fa-sign-in"></i>',
-            login=_('Login')
-        ),
-        order=1,
-        id='login-menu',
-    )
-
-    item = current_menu.submenu('main.profile.professional')
-    rero_register(
-        item,
-        endpoint='rero_ils.professional',
-        visible_when=lambda: current_librarian,
-        text='{icon} {professional}'.format(
-            icon='<i class="fa fa-briefcase"></i>',
-            professional=_('Professional interface')
-        ),
-        order=1,
-        id='professional-interface-menu',
-    )
-
-    item = current_menu.submenu('main.profile.logout')
-    rero_register(
-        item,
-        endpoint='security.logout',
-        endpoint_arguments_constructor=lambda: dict(
-            next='/{viewcode}'.format(viewcode=request.view_args.get(
-                'viewcode', current_app.config.get(
-                    'RERO_ILS_SEARCH_GLOBAL_VIEW_CODE'))
-            )
-        ),
-        visible_when=lambda: current_user.is_authenticated,
-        text='{icon} {logout}'.format(
-            icon='<i class="fa fa-sign-out"></i>',
-            logout=_('Logout')
-        ),
-        order=2,
-        id='logout-menu',
-    )
-
-    item = current_menu.submenu('main.profile.profile')
-    profile_endpoint = 'patrons.profile'
-    rero_register(
-        item,
-        endpoint=profile_endpoint,
-        visible_when=lambda: len(current_patrons) > 0,
-        text='{icon} {profile}'.format(
-            icon='<i class="fa fa-book"></i>',
-            profile=_('My Account')
-        ),
-        order=1,
-        id='profile-menu',
-    )
-
-    item = current_menu.submenu('main.profile.edit_profile')
-    rero_register(
-        item,
-        endpoint='invenio_userprofiles.profile',
-        visible_when=lambda: current_user.is_authenticated,
-        text='{icon} {profile}'.format(
-            icon='<i class="fa fa-user"></i>',
-            profile=_('Edit my profile')
-        ),
-        order=1,
-        id='profile-menu',
-    )
-
-    item = current_menu.submenu('main.profile.change_password')
-    rero_register(
-        item,
-        endpoint='security.change_password',
-        visible_when=lambda: current_user.is_authenticated,
-        text='{icon} {profile}'.format(
-            icon='<i class="fa fa-lock"></i>',
-            profile=_('Change password')
-        ),
-        order=1,
-        id='profile-menu',
-    )
-
-    # Endpoint for:
-    # Application: invenio_oauth2server_settings.index
-    # Security: invenio_accounts.security
-
-    item = current_menu.submenu('main.profile.signup')
-    rero_register(
-        item,
-        endpoint='security.register',
-        visible_when=lambda: not current_user.is_authenticated,
-        text='{icon} {signup}'.format(
-            icon='<i class="fa fa-user-plus"></i>',
-            signup=_('Sign Up')
-        ),
-        order=2,
-        id='signup-menu',
-    )
-
-
-@blueprint.before_app_request
+@blueprint.before_app_first_request
 def init_menu():
     """Create the header menus."""
     if (request.endpoint not in ['patrons.logged_user',
